backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
pymupdf
pypdf
reportlab
python-docx
mammoth
pdf2image
numpy
pillow
pandas
openpyxl
xlrd
pytesseract
ffmpeg-python
pydub
supabase
openai
stripe
psutil
httpx
passlib[bcrypt]
psycopg2-binary